Move popupUrl to file
